Added Firebase Anon Auth and Placing Ticket on Screen
Added a rough implementation of placing the ticket object on the screen. This is only so that when I place them on the server and pull them I have a way to test visually that I have grabbed the right ticket. The firebase anon sign in is done when the user hits the login button on the login page.