Added Saving Tickets to Server
I have added the ability to add the tickets to the server. When the button at the bottom of the home page is pressed you will be able to create a new ticket with the text fields provided. I will next be working on a way to pull all the open tickets at launch and display them on the screen.