Added Changing Ticket State and Ticket UI Updates
I changed the ticket Ui to be an expanding card. Instead of overwhelming the Homepage UI it makes it seem more clean. There is now a change ticket state button once the cards are expanded. This will change the state to the opposite of what it is and move it to the correct place in the database. Currently there is no way to switch a ticket from closed to open. This is due to the closed tickets needing to be placed, after closed, on a new page. Therefore I have not enabled a way to change the state back until this is completed.