Added Admin, Closed Tickets, Large UI changes
I added actual Admin functionality to the project. When logged in as an admin you are able to change the states of the tickets. I also added a way to view closed tickets and reopen them if needed. I also changed our theme as it was not looking too hot when I added a side bar to the project.